ticket-*: explicitly add tslib
This fixes an error when building a production build after upgrading angular versions

This also removes the yarn dependency inside the package since it isn't needed nor should it be there in the first place, it gets installed in the docker file regardless. There is no reason angular should try to treeshake or bundle it.